refactor: The main file creating the getCustomData
