+++ src/templates/laravel/model
@@ -24,11 +24,11 @@ class <%= tableMasterSingularUCamelCase; %> extends Model
     protected $hidden = ['created_at', 'updated_at', 'deleted_at'];
 
     protected $casts = [ /* 'field_name' => 'field_type' */ ];
-
+    <% if (tableMasterForeignKeysAssoc.length) { %>
     protected $with = [<% tableMasterForeignKeysAssoc.forEach(function(tbl, index){
     %>'<%= tbl.columnNameSingular; %>'<%= index < tableMasterForeignKeysAssoc.length-1 ? ',' : '' %><%
     }); %>];
-
+    <%}%>
     <% tableMasterForeignKeysAssoc.forEach(function(tbl){
     %>public function <%= tbl.columnNameSingular; %>()
     {
